requirements.txt
appdirs==1.4.3
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
black==19.3b0
boto3==1.9.158
botocore==1.12.158
certifi==2018.11.29
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.4
coverage==4.5.3
cryptography==3.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-admin-rangefilter==0.6.4
django-allauth==0.38.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-material-admin==1.7.16
django-rest-auth==0.9.3
django-storages==1.7.1
django-taggit==1.3.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
isort==4.3.4
jmespath==0.9.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient @ file:///C:/Users/User/Desktop/RupseOnline/rupseonline/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
oauthlib==3.0.1
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-decouple==3.1
python-magic==0.4.22
python3-openid==3.1.0
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.4.1
stripe==2.23.0
toml==0.10.0
typed-ast==1.4.2
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.0
